Sbusの設定をUsetSettingに移動(大会後のバージョンについて Fixes #23)
